Move http-based dependencies of chain_client behind rpc feature flag
So the crate can be imported and used in zkVM guest code.
