+++ rust/services/chain/client/Cargo.toml
@@ -6,19 +6,23 @@ edition = "2021"
 [dependencies]
 alloy-primitives = { workspace = true }
 async-trait = { workspace = true }
-block_trie = { workspace = true }
 chain_common = { workspace = true }
-chain_server = { workspace = true }
 derive-new = { workspace = true }
 futures = { workspace = true }
-mpt = { workspace = true }
 parking_lot = { workspace = true }
-provider = { workspace = true }
-serde_json = { workspace = true }
-server_utils = { workspace = true }
 thiserror = { workspace = true }
-tracing = { workspace = true }
+
+# RPC dependencies (disabled for guest)
+chain_server = { workspace = true, optional = true }
+mpt = { workspace = true, optional = true }
+serde_json = { workspace = true, optional = true }
+server_utils = { workspace = true, optional = true }
+tracing = { workspace = true, optional = true }
 
 [dev-dependencies]
 anyhow = { workspace = true }
+block_trie = { workspace = true }
 tokio = { workspace = true, features = ["rt", "macros"] }
+
+[features]
+rpc = ["chain_server", "mpt", "serde_json", "server_utils", "tracing"]
